requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain & AI
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.7
tiktoken==0.5.2

# Vector Stores & Embeddings
pgvector==0.2.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3

# Database & Storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
opentelemetry-instrumentation-openai==0.2.0
opentelemetry-exporter-prometheus==1.12.0rc1
prometheus-client==0.19.0

# Utilities
pyyaml==6.0.1
jinja2==3.1.2
redis==5.0.1
httpx==0.25.2
tenacity==8.2.3
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
